Help schedule find log files
This ought to resolve a recurring crontab error caused by not finding the app log files under `/var/log`. Note that I could not connect to `preservation-robots1-{stage,prod}` to troubleshoot this further, so I am putting this PR with less confidence than I would like. (And I have begun a discussion in Slack about not being able to connect.)