Add feature flag for transfer via scp
Fixes #417

This commit adds a feature flag in front of a new feature: running bag transfers via scp instead of tarpipe. Once we are sure this change is harmless, and ideally offers some benefit in the form of hogging less CPU cycles, we can remove the feature flag and make it the default behavior.